+++ riocore/cfggraph.py
@@ -1,4 +1,8 @@
+import json
 import graphviz
+
+from PyQt5.QtCore import Qt, QByteArray, QBuffer, QIODevice, QPointF
+from PyQt5.QtGui import QPixmap, QPainter, QPen, QColor, QPainterPath
 
 
 class ConfigGraph:
@@ -352,6 +356,119 @@ class ConfigGraph:
                     self.map[instance].append((int(begin[0]), int(begin[1]), int(end[0]), int(end[1])))
 
             return gAll.pipe()
+
+            """
+            jmap = json.loads(gAll.pipe(format="json").decode())
+            x, y, w, h = jmap["bb"].split(",")
+            pw = int(w)
+            ph = int(h)
+            pixmap = QPixmap(int(w) + 5, int(h) + 5)
+            pixmap.fill(Qt.black)
+            painter = QPainter(pixmap)
+            color = QColor("red")
+            pen = QPen(color, 1)
+            painter.setPen(pen)
+            # painter.drawLine(0, 0, pw, ph);
+
+            def hex_to_rgb(hex):
+                return tuple(int(hex[i : i + 2], 16) for i in (0, 2, 4))
+
+            for edge in jmap.get("edges", []):
+                for draw in edge.get("_draw_", {}):
+                    if draw.get("op") == "c":
+                        rgb = hex_to_rgb(draw["color"][1:])
+                        color.setRgb(*rgb)
+                        pen = QPen(color, 1)
+                        painter.setPen(pen)
+                    elif draw.get("op") == "b":
+                        x, y = draw["points"].pop(0)
+                        y = ph - y
+                        path = QPainterPath()
+                        path.moveTo(x, y)
+                        while len(draw["points"]) >= 3:
+                            cpoints = []
+                            x, y = draw["points"].pop(0)
+                            y = ph - y
+                            cpoints.append(float(x))
+                            cpoints.append(float(y))
+                            x, y = draw["points"].pop(0)
+                            y = ph - y
+                            cpoints.append(float(x))
+                            cpoints.append(float(y))
+                            x, y = draw["points"].pop(0)
+                            y = ph - y
+                            cpoints.append(float(x))
+                            cpoints.append(float(y))
+                            path.cubicTo(*cpoints)
+
+                        painter.drawPath(path)
+
+            color = QColor("white")
+            pen = QPen(color, 1)
+            painter.setPen(pen)
+
+            for part in jmap["objects"]:
+                # print(part)
+                url = part.get("URL")
+                for ldraw in part.get("_draw_"):
+                    # if draw.get("op") == "c":
+                    #    rgb = hex_to_rgb(draw["color"][1:])
+                    #    color.setRgb(*rgb)
+                    #    pen = QPen(color, 1)
+                    #    painter.setPen(pen)
+                    if ldraw["op"] == "L":
+                        x1, y1 = ldraw["points"][0]
+                        x2, y2 = ldraw["points"][1]
+
+                        x1 = int(float(x1))
+                        y1 = ph - int(float(y1))
+                        x2 = int(float(x2))
+                        y2 = ph - int(float(y2))
+                        w = x2 - x1
+                        h = y2 - y1
+
+                        # painter.drawRect(x1, y1, w, h)
+                        painter.drawLine(x1, y1, x2, y2)
+
+                for ldraw in part.get("_ldraw_"):
+                    # if draw.get("op") == "c":
+                    #    rgb = hex_to_rgb(draw["color"][1:])
+                    #    color.setRgb(*rgb)
+                    #    pen = QPen(color, 1)
+                    #    painter.setPen(pen)
+                    if ldraw["op"] == "T":
+                        x, y = ldraw["pt"]
+                        y = ph - y
+                        w = ldraw["width"]
+                        if ldraw["align"] == "c":
+                            x -= w / 2
+                            y += 3
+                        painter.drawText(QPointF(x, y), ldraw["text"])
+
+                for rect in part.get("rects").split(" "):
+                    x1, y1, x2, y2 = rect.split(",")
+                    x1 = int(float(x1))
+                    y1 = ph - int(float(y1))
+                    x2 = int(float(x2))
+                    y2 = ph - int(float(y2))
+                    w = x2 - x1
+                    h = y2 - y1
+
+                    painter.drawRect(x1, y1, w, h)
+                    if url:
+                        instance = url.split(":")[1].replace("#", " ")
+                        if instance not in self.map:
+                            self.map[instance] = []
+                        self.map[instance].append((int(float(x1)), int(float(y2)), int(float(x2)), int(float(y1))))
+
+            painter.end()
+
+            png_bytes = QByteArray()
+            png_buffer = QBuffer(png_bytes)
+            png_buffer.open(QIODevice.WriteOnly)
+            pixmap.save(png_buffer, "PNG")
+            return png_bytes
+            """
 
         except Exception as error:
             print(f"ERROR(overview): {error}")